requirements.txt
SQLAlchemy>=2.0.0
pymysql>=1.1.0
requests>=2.31.0
Flask>=3.0.0
pywinauto>=0.6.8
PyAutoGUI>=0.9.54
opencv-python-headless>=4.8.0
pillow>=10.0.0
pytesseract>=0.3.10
pyperclip>=1.8.2
